pyproject.toml
[project]
name = "fastapi-adk-practice"
version = "0.1.0"
description = "FastAPI and Google ADK practice project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "bcrypt==4.0.1",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
